+++ dzVents/documentation/history.md
@@ -1,3 +1,8 @@
+[2.4.27]
+- Add protect attribute for devices / scenes and groups
+- Add methods protectOn and protectOff for devices / scenes and groups
+- Add functions rpad, lpad, mpad, zpad, numDecimals in utils
+
 [2.4.26]
 - Add Smoke detector